backend/requirements.txt
fastapi
uvicorn
requests
gitpython
python-dotenv
pytest